==> src/01-data-prep/06-calculate-ward-areas.ipynb <==
import geopandas as gpd
import pandas as pd

# Path to the uploaded GeoJSON file
geojson_file_path = '../../initial-csv-data/Wards-pre-2013.json'
# Path to the new CSV file
new_csv_file_path = '../../working-csv-data/06-wards-area.csv'

# Load the GeoJSON file
wards_df = gpd.read_file(geojson_file_path)

# Convert to a projected CRS suitable for Northern Ireland (British National Grid - EPSG:27700)
# This line of code changes the CRS of the GeoDataFrame 'wards_df' from its current CRS to the British National Grid (EPSG:27700).
# A CRS defines how the two-dimensional, flat map relates to real places on the earth. The British National Grid is a coordinate system
# used in Great Britain, suitable for high-accuracy local mapping and analysis in Northern Ireland.
# The 'to_crs' method is used for this transformation, ensuring that spatial operations such as area calculation are more accurate,
# as they consider the curvature of the earth and variations in surface measurements due to projection.
wards_df = wards_df.to_crs(epsg=27700)

# Calculate the area in square kilometers
# After transforming the CRS of 'wards_df' to a projected system (EPSG:27700), we can now accurately calculate the area of each geometry.
# The '.area' attribute of a GeoDataFrame calculates the area of each geometry within it. However, this calculation is done in the CRS's
# units, which, for EPSG:27700, are square meters.
# To convert the area from square meters to square kilometers, we divide each area by 10^6 (since 1 square kilometer equals 1,000,000 square meters).
# The result of this calculation is stored in a new column in 'wards_df' called 'area_sqkm', representing the area of each ward in square kilometers.
wards_df['area_sqkm'] = wards_df.area / 10**6

# Create a new DataFrame with ward names and their areas
# Assuming 'WARDS' is the correct column name for ward names
wards_area_df = wards_df[['WARDS', 'area_sqkm']]


# Save to CSV
wards_area_df.to_csv(new_csv_file_path, index=False)

print(f'CSV file created at {new_csv_file_path}')
